Modularise the JS
- Pull data into json and into hugo data folder. Next step, static build all the things.
- Decompose common.js into named modules with a single purpose, to ease maintenance.
- Fix process to create page specific and common JS bundles.
- Add build:* and watch* scripts to package.json
