+++ content/index.html
@@ -219,14 +219,14 @@
 <script type="text/javascript" src="/js/index.js"></script>
 
 <!--Adobe Edge Runtime-->
-<script type="text/javascript" charset="utf-8" src="/js/lib/edge6.0.0.js"></script>
+<script type="text/javascript" charset="utf-8" src="/js/vendor/edge6.0.0.js"></script>
 <style>
     .edgeLoad-EDGE-1720491910 {
         visibility: hidden;
     }
 </style>
 <script>
-    AdobeEdge.loadComposition('js/lib/libp2p', 'EDGE-1720491910', {
+    AdobeEdge.loadComposition('js/vendor/libp2p', 'EDGE-1720491910', {
         scaleToFit: "horizontal",
         centerStage: "horizontal",
         minW: "0px",
