+++ content/media.html
@@ -52,6 +52,7 @@
   {{ partial "footer.html" . }}
 
   <script type="text/javascript" src="/js/common.js"></script>
+  <script type="text/javascript" src="/js/media.js"></script>
 
 </body>
 </html>
