+++ data/bundles.json
@@ -1,0 +1,243 @@
+{
+  "Bundles": {
+    "Browser JS": {
+      "status": "live",
+      "image": "../img/logo_1.png",
+      "github": "https://github.com/ipfs/js-libp2p-ipfs-browser",
+      "categories": {
+        "Transport": {
+          "modules": {
+            "libp2p-websockets": {
+              "status": "Done",
+              "url": "https://github.com/libp2p/js-libp2p-websockets"
+            },
+            "libp2p-webrtc-star": {
+              "status": "Done",
+              "url": "https://github.com/libp2p/js-libp2p-webrtc-star"
+            }
+          }
+        },
+        "Stream Muxer": {
+          "modules": {
+            "libp2p-multiplex": {
+              "status": "Done",
+              "url": "https://github.com/libp2p/js-libp2p-multiplex"
+            },
+            "libp2p-spdy": {
+              "status": "Done",
+              "url": "https://github.com/libp2p/js-libp2p-spdy"
+            }
+          }
+        },
+        "Crypto Channels": {
+          "modules": {
+            "libp2p-secio": {
+              "status": "Done",
+              "url": "https://github.com/libp2p/js-libp2p-secio"
+            }
+          }
+        },
+        "Peer Routing": {
+          "modules": {
+            "libp2p-kad-dht": {
+              "status": "unstable",
+              "url": "https://github.com/libp2p/js-libp2p-kad-dht"
+            }
+          }
+        },
+        "Record Stores": {
+          "modules": {
+            "record": {
+              "status": "usable",
+              "url": "https://github.com/libp2p/js-libp2p-record"
+            }
+          }
+        },
+        "Discovery": {
+          "modules": {
+            "libp2p-webrtc-star": {
+              "status": "done",
+              "url": "https://github.com/libp2p/js-libp2p-webrtc-star"
+            },
+            "libp2p-railing": {
+              "status": "done",
+              "url": "https://github.com/libp2p/js-libp2p-railing"
+            }
+          }
+        }
+      }
+    },
+    "Node JS": {
+      "status": "live",
+      "image": "../img/logo_2.png",
+      "github": "https://github.com/ipfs/js-libp2p-ipfs-nodejs",
+      "categories": {
+        "Transport": {
+          "modules": {
+            "libp2p-tcp": {
+              "status": "Done",
+              "url": "https://github.com/libp2p/js-libp2p-tcp"
+            },
+            "libp2p-websockets": {
+              "status": "Done",
+              "url": "https://github.com/libp2p/js-libp2p-websockets"
+            },
+            "libp2p-webrtc-star": {
+              "status": "Done",
+              "url": "https://github.com/libp2p/js-libp2p-webrtc-star"
+            }
+          }
+        },
+        "Stream Muxer": {
+          "modules": {
+            "libp2p-multiplex": {
+              "status": "Done",
+              "url": "https://github.com/libp2p/js-libp2p-multiplex"
+            },
+            "libp2p-spdy": {
+              "status": "Done",
+              "url": "https://github.com/libp2p/js-libp2p-spdy"
+            }
+          }
+        },
+        "Crypto Channels": {
+          "modules": {
+            "libp2p-secio": {
+              "status": "Done",
+              "url": "https://github.com/libp2p/js-libp2p-secio"
+            }
+          }
+        },
+        "Peer Routing": {
+          "modules": {
+            "libp2p-kad-dht": {
+              "status": "unstable",
+              "url": "https://github.com/libp2p/js-libp2p-kad-dht"
+            }
+          }
+        },
+        "Record Stores": {
+          "modules": {
+            "record": {
+              "status": "usable",
+              "url": "https://github.com/libp2p/js-libp2p-record"
+            }
+          }
+        },
+        "Discovery": {
+          "modules": {
+            "libp2p-mdns": {
+              "status": "done",
+              "url": "https://github.com/libp2p/js-libp2p-mdns"
+            },
+            "libp2p-webrtc-star": {
+              "status": "done",
+              "url": "https://github.com/libp2p/js-libp2p-webrtc-star"
+            },
+            "libp2p-railing": {
+              "status": "unstable",
+              "url": "https://github.com/libp2p/js-libp2p-railing"
+            }
+          }
+        }
+      }
+    },
+    "Golang": {
+      "status": "live",
+      "image": "../img/logo_3.png",
+      "github": "https://github.com/libp2p/go-libp2p",
+      "categories": {
+        "Transport": {
+          "modules": {
+            "go-tcp-transport": {
+              "status": "Done",
+              "url": "https://github.com/libp2p/go-tcp-transport"
+            },
+            "go-ws-transport": {
+              "status": "Done",
+              "url": "https://github.com/libp2p/go-ws-transport"
+            }
+          }
+        },
+        "Stream Muxer": {
+          "modules": {
+            "libp2p-yamux": {
+              "status": "Done",
+              "url": ""
+            },
+            "libp2p-multiplex": {
+              "status": "Done",
+              "url": ""
+            }
+          }
+        },
+        "Crypto Channels": {
+          "modules": {
+            "libp2p-secio": {
+              "status": "Done",
+              "url": "https://github.com/libp2p/go-libp2p-secio"
+            }
+          }
+        },
+        "Peer Routing": {
+          "modules": {
+            "libp2p-kad-dht": {
+              "status": "Done",
+              "url": "https://github.com/libp2p/go-libp2p-kad-dht"
+            }
+          }
+        },
+        "Record Stores": {
+          "modules": {
+            "libp2p-kad-dht": {
+              "status": "Done",
+              "url": "https://github.com/libp2p/go-libp2p-kad-dht"
+            }
+          }
+        },
+        "NAT Traversal": {
+          "modules": {
+            "libp2p-nat": {
+              "status": "Done",
+              "url": "https://github.com/libp2p/go-libp2p-nat"
+            }
+          }
+        },
+        "Discovery": {
+          "modules": {
+            "libp2p-mdns": {
+              "status": "done",
+              "url": "https://github.com/libp2p/go-libp2p/blob/master/p2p/discovery/mdns.go"
+            },
+            "libp2p-railing": {
+              "status": "unstable",
+              "url": ""
+            }
+          }
+        }
+      }
+    },
+    "Haskell": {
+      "status": "coming-soon",
+      "image": "../img/logo_4.png",
+      "github": ""
+    },
+    "Java": {
+      "status": "coming-soon",
+      "image": "../img/logo_java.png",
+      "nav-img": "../img/logo_5.png",
+      "github": ""
+    },
+    "Python": {
+      "status": "coming-soon",
+      "image": "../img/logo_6.png",
+      "nav-img": "../img/logo_6_.png",
+      "github": ""
+    },
+    "Rust": {
+      "status": "coming-soon",
+      "image": "../img/logo_7.png",
+      "github": ""
+    }
+  }
+}
