--- js/data/implementations_data.js
@@ -1,338 +1,0 @@
-module.exports = {
-  'Transports': {
-    'libp2p-tcp': {
-      'Browser JS': {
-        status: 'Done',
-        url: 'https://github.com/libp2p/js-libp2p-tcp'
-      },
-      'Node.js': {
-        status: 'Done',
-        url: 'https://github.com/libp2p/js-libp2p-tcp'
-      },
-      'Go': {
-        status: 'Done',
-        url: 'https://github.com/libp2p/go-tcp-transport'
-      }
-    },
-    'libp2p-quic': {
-      'Go': {
-        status: 'Unstable',
-        url: 'https://github.com/marten-seemann/libp2p-quic-transport'
-      }
-    },
-    'libp2p-websockets': {
-      'Browser JS': {
-        status: 'Done',
-        url: 'https://github.com/libp2p/js-libp2p-websockets'
-      },
-      'Node.js': {
-        status: 'Done',
-        url: 'https://github.com/libp2p/js-libp2p-websockets'
-      },
-      'Go': {
-        status: 'Done',
-        url: 'https://github.com/libp2p/go-ws-transport'
-      }
-    },
-    'libp2p-webrtc-star': {
-      'Browser JS': {
-        status: 'Done',
-        url: 'https://github.com/libp2p/js-libp2p-webrtc-star'
-      },
-      'Node.js': {
-        status: 'Done',
-        url: 'https://github.com/libp2p/js-libp2p-webrtc-star'
-      }
-    },
-    'libp2p-webrtc-direct': {
-      'Browser JS': {
-        status: 'Done',
-        url: 'https://github.com/libp2p/js-libp2p-webrtc-direct'
-      },
-      'Node.js': {
-        status: 'Done',
-        url: 'https://github.com/libp2p/js-libp2p-webrtc-direct'
-      }
-    },
-    'libp2p-udp': {
-      'Node.js': {
-        status: 'Unstable',
-        url: 'https://github.com/libp2p/js-libp2p-udp'
-      },
-      'Go': {
-        status: 'Unstable',
-        url: 'https://github.com/libp2p/go-udp-transport'
-      }
-    },
-    'libp2p-utp': {
-      'Node.js': {
-        status: 'Unstable',
-        url: 'https://github.com/libp2p/js-libp2p-utp'
-      },
-      'Go': {
-        status: 'Usable',
-        url: 'https://github.com/libp2p/go-utp-transport'
-      }
-    }
-  },
-  'stream-muxers': {
-    'libp2p-spdy': {
-      'Browser JS': {
-        status: 'Done',
-        url: 'https://github.com/libp2p/js-libp2p-spdy'
-      },
-      'Node.js': {
-        status: 'Done',
-        url: 'https://github.com/libp2p/js-libp2p-spdy'
-      },
-      'Go': {
-        status: 'Unstable',
-        url: 'https://github.com/docker/spdystream'
-      }
-    },
-    'libp2p-multiplex': {
-      'Browser JS': {
-        status: 'Done',
-        url: 'https://github.com/libp2p/js-libp2p-multiplex'
-      },
-      'Node.js': {
-        status: 'Done',
-        url: 'https://github.com/libp2p/js-libp2p-multiplex'
-      },
-      'Go': {
-        status: 'Done',
-        url: 'https://github.com/whyrusleeping/go-smux-multiplex'
-      }
-    },
-    'libp2p-yamux': {
-      'Go': {
-        status: 'Done',
-        url: 'https://github.com/hashicorp/yamux'
-      }
-    }
-  },
-  'crypto-channels': {
-    'libp2p-secio': {
-      'Browser JS': {
-        status: 'Done',
-        url: 'https://github.com/libp2p/js-libp2p-secio'
-      },
-      'Node.js': {
-        status: 'Done',
-        url: 'https://github.com/libp2p/js-libp2p-secio'
-      },
-      'Go': {
-        status: 'Done',
-        url: 'https://github.com/libp2p/go-libp2p-secio'
-      }
-    }
-  },
-  'connection-upgrades': {
-    'libp2p-conn': {
-      'Browser JS': {
-        status: 'Done',
-        url: 'https://github.com/libp2p/interface-connection'
-      },
-      'Node.js': {
-        status: 'Done',
-        url: 'https://github.com/libp2p/interface-connection'
-      },
-      'Go': {
-        status: 'Done',
-        url: 'https://github.com/libp2p/go-libp2p-conn'
-      }
-    },
-    'half-closed-connection-upgrade': {
-      'Browser JS': {
-        status: 'Unstable',
-        url: 'https://github.com/libp2p/js-libp2p-half-closed-connection-upgrade'
-      },
-      'Node.js': {
-        status: 'Usable',
-        url: 'https://github.com/libp2p/js-libp2p-half-closed-connection-upgrade'
-      }
-    }
-  },
-  'peer-routing': {
-    'libp2p-kad-dht': {
-      'Browser JS': {
-        status: 'Unstable',
-        url: 'https://github.com/libp2p/js-libp2p-dht'
-      },
-      'Node.js': {
-        status: 'Unstable',
-        url: 'https://github.com/libp2p/js-libp2p-dht'
-      },
-      'Go': {
-        status: 'Done',
-        url: 'https://github.com/libp2p/go-libp2p-kad-dht'
-      }
-    }
-  },
-  'record-stores': {
-    'record': {
-      'Browser JS': {
-        status: 'Usable',
-        url: 'https://github.com/libp2p/js-libp2p-record'
-      },
-      'Node.js': {
-        status: 'Usable',
-        url: 'https://github.com/libp2p/js-libp2p-record'
-      },
-      'Go': {
-        status: 'Done',
-        url: 'https://github.com/libp2p/go-libp2p-record'
-      },
-      'C#': {
-        status: 'Unstable',
-        url: 'https://github.com/libp2p/go-libp2p-record'
-      }
-    }
-  },
-  'nat-traversal': {
-    'libp2p-nat': {
-      'Go': {
-        status: 'Unstable',
-        url: 'https://github.com/libp2p/go-libp2p-nat'
-      }
-    }
-  },
-  'discovery': {
-    'bootstrap': {
-      'Browser JS': {
-        status: 'Usable',
-        url: 'https://github.com/libp2p/js-libp2p-railing'
-      },
-      'Node.js': {
-        status: 'Done',
-        url: 'https://github.com/libp2p/js-libp2p-railing'
-      },
-      'Go': {
-        status: 'Done',
-        url: 'https://github.com/ipfs/go-ipfs/blob/master/core/bootstrap.go'
-      }
-    },
-    'random-walk': {
-      'Browser JS': {
-        status: 'Unstable',
-        url: 'https://github.com/libp2p/js-libp2p-random-walk'
-      },
-      'Node.js': {
-        status: 'Unstable',
-        url: 'https://github.com/libp2p/js-libp2p-random-walk'
-      },
-      'Go': {
-        status: 'Done',
-        url: 'https://github.com/libp2p/go-libp2p-kad-dht'
-      }
-    },
-    'mdns-discovery': {
-      'Node.js': {
-        status: 'Done',
-        url: 'https://github.com/libp2p/js-libp2p-mdns'
-      },
-      'Go': {
-        status: 'Done',
-        url: 'https://github.com/libp2p/go-libp2p/blob/master/p2p/discovery/mdns.go'
-      }
-    }
-  },
-  'utils': {
-    'crypto': {
-      'Browser JS': {
-        status: 'Done',
-        url: 'https://github.com/libp2p/js-libp2p-crypto'
-      },
-      'Node.js': {
-        status: 'Done',
-        url: 'https://github.com/libp2p/js-libp2p-crypto'
-      },
-      'Go': {
-        status: 'Done',
-        url: 'https://github.com/libp2p/go-libp2p-crypto'
-      }
-    },
-    'libp2p-ping': {
-      'Browser JS': {
-        status: 'Done',
-        url: 'https://github.com/libp2p/js-libp2p-ping'
-      },
-      'Node.js': {
-        status: 'Done',
-        url: 'https://github.com/libp2p/js-libp2p-ping'
-      },
-      'Go': {
-        status: 'Done',
-        url: 'https://github.com/libp2p/go-libp2p/p2p/protocol/ping'
-      }
-    },
-    'libp2p-peer-id': {
-      'Browser JS': {
-        status: 'Done',
-        url: 'https://github.com/libp2p/js-peer-id'
-      },
-      'Node.js': {
-        status: 'Done',
-        url: 'https://github.com/libp2p/js-peer-id'
-      }
-    },
-    'libp2p-peer-info': {
-      'Browser JS': {
-        status: 'Done',
-        url: 'https://github.com/libp2p/js-peer-info'
-      },
-      'Node.js': {
-        status: 'Done',
-        url: 'https://github.com/libp2p/js-peer-info'
-      },
-      'Go': {
-        status: 'Done',
-        url: 'https://github.com/libp2p/go-libp2p-peer'
-      }
-    },
-    'libp2p-peer-book': {
-      'Browser JS': {
-        status: 'Done',
-        url: 'https://github.com/libp2p/js-peer-book'
-      },
-      'Node.js': {
-        status: 'Done',
-        url: 'https://github.com/libp2p/js-peer-book'
-      },
-      'Go': {
-        status: 'Done',
-        url: 'https://github.com/libp2p/go-libp2p-peerstore'
-      }
-    },
-    'libp2p-swarm': {
-      'Browser JS': {
-        status: 'Done',
-        url: 'https://github.com/libp2p/js-libp2p-swarm'
-      },
-      'Node.js': {
-        status: 'Done',
-        url: 'https://github.com/libp2p/js-libp2p-swarm'
-      },
-      'Go': {
-        status: 'Done',
-        url: 'https://github.com/libp2p/go-libp2p/tree/master/p2p/host/basic'
-      }
-    }
-  },
-  'others': {
-    'libp2p-floodsub': {
-      'Browser JS': {
-        status: 'Usable',
-        url: 'https://github.com/libp2p/js-libp2p-floodsub'
-      },
-      'Node.js': {
-        status: 'Usable',
-        url: 'https://github.com/libp2p/js-libp2p-floodsub'
-      },
-      'Go': {
-        status: 'Done',
-        url: 'https://github.com/libp2p/go-floodsub'
-      }
-    }
-  }
-}
