+++ js/lib/scroll-to-hash.js
@@ -1,0 +1,9 @@
+var $ = require('jquery')
+
+module.exports = function scrollToHash () {
+  var hash = window.location.hash
+  if (hash.length) {
+    var $scrollTo = $(hash)
+    $('html, body').animate({scrollTop: $scrollTo.offset().top - 100}, 500, 'swing')
+  }
+}
