+++ js/media.js
@@ -1,0 +1,3 @@
+var initPage = require('./lib/init-page')
+
+initPage()
